Cast offset to long long

+++ librz/magic/magic-test.c
@@ -1344,9 +1344,9 @@ magic_test_line(RzMagicLine *ml, RzMagicState *ms) {
 	if (ml->mimetype != NULL)
 		ms->mimetype = ml->mimetype;
 
-	magic_warn(ml, "test %s/%c matched at offset %ld (now %zu): "
+	magic_warn(ml, "test %s/%c matched at offset %lld (now %zu): "
 		       "'%s'",
-		ml->type_string, ml->test_operator, offset,
+		ml->type_string, ml->test_operator, (long long)offset,
 		ms->offset, ml->result == NULL ? "" : ml->result);
 
 	if (ml->type == MAGIC_TYPE_USE) {
@@ -1358,7 +1358,7 @@ magic_test_line(RzMagicLine *ml, RzMagicState *ms) {
 			magic_warn(ml, "no name found for use %s", ml->name);
 			return (0);
 		}
-		magic_warn(ml, "use %s at offset %ld", ml->name, offset);
+		magic_warn(ml, "use %s at offset %lld", ml->name, (long long)offset);
 		magic_test_children(named, ms, offset, *ml->name == '^');
 	}
